requirements.txt
requests
joblibs